factor out more of smart_cards_screen
