+++ libs/ui/src/smart_cards_screen.tsx
@@ -1,10 +1,23 @@
 import styled from 'styled-components';
-import { assert, throwIllegalValue } from '@votingworks/basics';
+import {
+  assert,
+  deepEqual,
+  Result,
+  throwIllegalValue,
+} from '@votingworks/basics';
 import { format, hyphenatePin } from '@votingworks/utils';
-import { Election, UserWithCard } from '@votingworks/types';
-import React from 'react';
+import {
+  constructElectionKey,
+  DippedSmartCardAuth,
+  Election,
+  ElectionDefinition,
+  SystemSettings,
+  UserWithCard,
+} from '@votingworks/types';
+import React, { useState } from 'react';
+import { useMutation } from '@tanstack/react-query';
 import { Callout } from './callout';
-import { H2, H3, P } from './typography';
+import { H1, H2, H3, P } from './typography';
 import { Button } from './button';
 import { Modal } from './modal';
 import { Icons } from './icons';
@@ -19,7 +32,7 @@ export const SmartCardsScreenButtonList = styled.div`
 `;
 
 function toLowerCaseExceptFirst(str: string): string {
-  // istanbul ignore next
+  /* istanbul ignore next - @preserve */
   if (str.length === 0) {
     return str;
   }
@@ -49,6 +62,22 @@ const CardActions = styled.div`
   display: flex;
   flex-direction: column;
 `;
+
+export function ElectionInfo({
+  election,
+}: {
+  election: Election;
+}): JSX.Element {
+  return (
+    <React.Fragment>
+      {election.title}
+      <br />
+      {format.localeWeekdayAndDate(
+        election.date.toMidnightDatetimeWithSystemTimezone()
+      )}
+    </React.Fragment>
+  );
+}
 
 export function InsertCardPrompt({
   cardStatus,
@@ -197,22 +226,6 @@ export function ActionResultCallout({
   }
 }
 
-export function ElectionInfo({
-  election,
-}: {
-  election: Election;
-}): JSX.Element {
-  return (
-    <React.Fragment>
-      {election.title}
-      <br />
-      {format.localeWeekdayAndDate(
-        election.date.toMidnightDatetimeWithSystemTimezone()
-      )}
-    </React.Fragment>
-  );
-}
-
 interface ConfirmSystemAdminCardAction {
   actionType: 'Program' | 'PinReset';
   doAction: VoidFunction;
@@ -282,5 +295,255 @@ export function ConfirmSystemAdminCardActionModal({
         </React.Fragment>
       }
     />
+  );
+}
+
+export interface CardProgrammingApiClient {
+  programCard: (input: {
+    userRole: 'system_administrator' | 'election_manager' | 'poll_worker';
+  }) => Promise<
+    Result<
+      {
+        pin?: string;
+      },
+      Error
+    >
+  >;
+  unprogramCard: () => Promise<Result<void, Error>>;
+}
+
+export function CardDetailsAndActions({
+  card,
+  systemSettings,
+  electionDefinition,
+  apiClient,
+}: {
+  card: DippedSmartCardAuth.ProgrammableCardReady;
+  systemSettings: SystemSettings;
+  electionDefinition: ElectionDefinition;
+  apiClient: CardProgrammingApiClient;
+}): JSX.Element {
+  const [actionResult, setActionResult] = useState<SmartCardActionResult>();
+  const [confirmSystemAdminCardAction, setConfirmSystemAdminCardAction] =
+    useState<ConfirmSystemAdminCardAction>();
+
+  const programCardMutation = useMutation(
+    (input: {
+      userRole: 'system_administrator' | 'election_manager' | 'poll_worker';
+    }) => apiClient.programCard(input)
+  );
+
+  const unprogramCardMutation = useMutation(() => apiClient.unprogramCard());
+
+  function programCard(role: CardRole) {
+    assert(role !== 'vendor');
+
+    programCardMutation.mutate(
+      { userRole: role },
+      {
+        onSuccess: (result) => {
+          setActionResult({
+            action: 'Program',
+            newPin: result.ok()?.pin,
+            role,
+            status: result.isOk() ? 'Success' : 'Error',
+          });
+        },
+      }
+    );
+  }
+
+  function unprogramCard(role: CardRole) {
+    unprogramCardMutation.mutate(undefined, {
+      onSuccess: (result) => {
+        setActionResult({
+          action: 'Unprogram',
+          role,
+          status: result.isOk() ? 'Success' : 'Error',
+        });
+      },
+    });
+  }
+
+  function resetCardPin(role: CardRole) {
+    assert(role !== 'vendor');
+    programCardMutation.mutate(
+      { userRole: role },
+      {
+        onSuccess: (result) => {
+          setActionResult({
+            action: 'PinReset',
+            newPin: result.ok()?.pin,
+            role,
+            status: result.isOk() ? 'Success' : 'Error',
+          });
+        },
+      }
+    );
+  }
+
+  const { programmedUser } = card;
+  const { role } = programmedUser ?? {};
+
+  const doesCardElectionMatchMachineElection =
+    electionDefinition &&
+    programmedUser &&
+    (programmedUser.role === 'election_manager' ||
+      programmedUser.role === 'poll_worker') &&
+    deepEqual(
+      programmedUser.electionKey,
+      constructElectionKey(electionDefinition.election)
+    );
+
+  const electionInfo = doesCardElectionMatchMachineElection ? (
+    <ElectionInfo election={electionDefinition.election} />
+  ) : (
+    'Unknown Election'
+  );
+
+  // Don't allow unprogramming system administrator cards to ensure election officials don't get
+  // accidentally locked out.
+  const unprogramAllowed =
+    role === 'election_manager' || role === 'poll_worker';
+  // Disable unprogramming when there's no election definition on
+  // the machine since cards can't be programmed in this state
+  const unprogramDisabled = !electionDefinition;
+  const unprogramVariant = doesCardElectionMatchMachineElection
+    ? 'danger'
+    : 'primary';
+
+  const resetPinAllowed =
+    role === 'system_administrator' ||
+    role === 'election_manager' ||
+    (systemSettings.auth.arePollWorkerCardPinsEnabled &&
+      role === 'poll_worker');
+  // Because PIN resetting completely reprograms the card under the hood, we also need the
+  // relevant election definition to be loaded for election manager and poll worker cards, so
+  // that we can write the proper election key
+  const resetPinDisabled = !(
+    role === 'system_administrator' || doesCardElectionMatchMachineElection
+  );
+
+  const createElectionCardsDisabled = !electionDefinition;
+
+  const actionInProgress =
+    unprogramCardMutation.isLoading || programCardMutation.isLoading;
+
+  return (
+    <React.Fragment>
+      <CardIllustration inserted active={actionInProgress}>
+        <div style={{ width: '100%' }}>
+          <H1>{role ? prettyRoles[role] : 'Blank'} Card</H1>
+          {programmedUser && role !== 'system_administrator' && (
+            <P>{electionInfo}</P>
+          )}
+        </div>
+        {actionInProgress && (
+          <Icons.Loading color="primary" style={{ height: '5rem' }} />
+        )}
+      </CardIllustration>
+      <div style={{ flexGrow: 1 }}>
+        {actionResult && (
+          <div style={{ padding: '1rem' }}>
+            <ActionResultCallout result={actionResult} />
+          </div>
+        )}
+        {programmedUser ? (
+          // After a successful action, hide the modify card actions to keep
+          // user focus on the success message.
+          // Also show no actions section for vendor cards.
+          !(actionResult?.status === 'Success' || role === 'vendor') && (
+            <CardActions>
+              {unprogramAllowed && unprogramDisabled && (
+                <Callout color="warning" icon="Info">
+                  Configure VxAdmin with an election package to enable modifying
+                  cards.
+                </Callout>
+              )}
+              <H2>Modify Card</H2>
+              <SmartCardsScreenButtonList>
+                {unprogramAllowed && (
+                  <Button
+                    onPress={() => unprogramCard(role)}
+                    disabled={unprogramDisabled || actionInProgress}
+                    icon="Delete"
+                    variant={unprogramVariant}
+                  >
+                    Unprogram Card
+                  </Button>
+                )}
+                {resetPinAllowed && (
+                  <Button
+                    onPress={() => {
+                      if (role === 'system_administrator') {
+                        setConfirmSystemAdminCardAction({
+                          actionType: 'PinReset',
+                          doAction: () => resetCardPin(role),
+                        });
+                      } else {
+                        resetCardPin(role);
+                      }
+                    }}
+                    disabled={resetPinDisabled || actionInProgress}
+                    variant={
+                      role === 'system_administrator' ? 'danger' : undefined
+                    }
+                    icon="RotateRight"
+                  >
+                    Reset Card PIN
+                  </Button>
+                )}
+              </SmartCardsScreenButtonList>
+            </CardActions>
+          )
+        ) : (
+          <CardActions>
+            {createElectionCardsDisabled && (
+              <Callout color="warning" icon="Info">
+                Configure VxAdmin with an election package to program election
+                manager and poll worker cards.
+              </Callout>
+            )}
+            <H2>Program New Card</H2>
+            <SmartCardsScreenButtonList style={{ alignItems: 'start' }}>
+              <Button
+                icon="Add"
+                variant="primary"
+                onPress={() => programCard('election_manager')}
+                disabled={createElectionCardsDisabled || actionInProgress}
+              >
+                Program Election Manager Card
+              </Button>
+              <Button
+                icon="Add"
+                variant="primary"
+                onPress={() => programCard('poll_worker')}
+                disabled={createElectionCardsDisabled || actionInProgress}
+              >
+                Program Poll Worker Card
+              </Button>
+              <Button
+                icon="Add"
+                onPress={() =>
+                  setConfirmSystemAdminCardAction({
+                    actionType: 'Program',
+                    doAction: () => programCard('system_administrator'),
+                  })
+                }
+                disabled={actionInProgress}
+              >
+                Program System Administrator Card
+              </Button>
+            </SmartCardsScreenButtonList>
+          </CardActions>
+        )}
+      </div>
+      {confirmSystemAdminCardAction && (
+        <ConfirmSystemAdminCardActionModal
+          {...confirmSystemAdminCardAction}
+          onClose={() => setConfirmSystemAdminCardAction(undefined)}
+        />
+      )}
+    </React.Fragment>
   );
 }
